+++ src/main/java/lotto/Application.java
@@ -1,11 +1,10 @@
 package lotto;
 
 import lotto.GameProcess.LottoProcess;
-import lotto.GameProcess.LottoProcessImpl;
 
 public class Application {
     public static void main(String[] args) {
-        LottoProcess lottoProcess = new LottoProcessImpl();
+        LottoProcess lottoProcess = new LottoProcess();
         lottoProcess.playGame();
     }
 }
